chore: refactor worker module
