+++ crates/ursa-gateway/src/admin/route/api/v1/post.rs
@@ -4,7 +4,7 @@ use axum::Extension;
 use hyper::StatusCode;
 use tokio::sync::RwLock;
 
-use crate::worker::cache::AdminCache;
+use crate::worker::cache::admin::AdminCache;
 
 pub async fn purge_cache_handler<Cache: AdminCache>(
     Extension(cache): Extension<Arc<RwLock<Cache>>>,
