+++ crates/ursa-gateway/src/worker/cache/admin.rs
@@ -1,0 +1,11 @@
+use super::Cache;
+
+pub trait AdminCache: Send + Sync + 'static {
+    fn purge(&mut self);
+}
+
+impl AdminCache for Cache {
+    fn purge(&mut self) {
+        self.tlrfu.purge();
+    }
+}
